+++ UpgradeWorld/commands/LocationRegister.cs
@@ -2,15 +2,20 @@ using System.Linq;
 using Service;
 
 namespace UpgradeWorld;
-public class LocationRegisterCommand {
-  public LocationRegisterCommand() {
-    CommandWrapper.Register("location_register", (int index, int subIndex) => {
+public class LocationRegisterCommand
+{
+  public LocationRegisterCommand()
+  {
+    CommandWrapper.Register("location_register", (int index, int subIndex) =>
+    {
       if (index == 0) return CommandWrapper.LocationIds();
       if (index == 1) return CommandWrapper.XZY("pos", "Coordinates for the position. If not given, player's position is used", subIndex);
       return null;
     });
-    new Terminal.ConsoleCommand("location_register", "[id] [x,z,y=player position] - Registers a location without placing it.", (args) => {
-      if (args.Length == 0) {
+    new Terminal.ConsoleCommand("location_register", "[id] [x,z,y=player position] - Registers a location without placing it.", (args) =>
+    {
+      if (args.Length == 0)
+      {
         Helper.Print(args.Context, "Error: Missing the location id.");
         return;
       }