+++ UpgradeWorld/commands/LocationsAdd.cs
@@ -6,18 +6,14 @@ public class LocationsAddCommand
 {
   public LocationsAddCommand()
   {
-    CommandWrapper.Register("locations_add", (int index) =>
-    {
-      if (index == 0) return CommandWrapper.LocationIds();
-      return FiltererParameters.Parameters;
-    }, FiltererParameters.GetAutoComplete());
+    CommandWrapper.Register("locations_add", (int index) => index == 0 ? CommandWrapper.LocationIds() : FiltererParameters.Parameters, FiltererParameters.GetAutoComplete());
     new Terminal.ConsoleCommand("locations_add", "[id1,id2,...] [...args] - Adds missing locations to generated areas.", (args) =>
     {
       LocationIdParameters pars = new(args);
       if (Helper.IsClient(args)) return;
       if (!pars.Valid(args.Context)) return;
-      Executor.AddOperation(new DistributeLocations(pars.Ids, pars.Start, pars.Chance, args.Context));
-      Executor.AddOperation(new SpawnLocations(args.Context, pars));
+      Executor.AddOperation(new DistributeLocations(args.Context, pars.Ids, pars.Start, pars.Chance));
+      Executor.AddOperation(new SpawnLocations(args.Context, pars.Ids, pars));
     }, optionsFetcher: () => ZoneSystem.instance.m_locations.Select(location => location.m_prefabName).ToList());
   }
 }