+++ UpgradeWorld/commands/Upgrade.cs
@@ -1,16 +1,17 @@
 using Service;
 
 namespace UpgradeWorld;
-public class UpgradeCommand {
-  public UpgradeCommand() {
-    CommandWrapper.Register("upgrade", (int index) => {
-      if (index == 0) return Upgrade.Types;
-      return FiltererParameters.Parameters;
-    }, FiltererParameters.GetAutoComplete());
-    new Terminal.ConsoleCommand("upgrade", "[operation] [...args] - Performs a predefined upgrade operation.", (args) => {
+public class UpgradeCommand
+{
+  public UpgradeCommand()
+  {
+    CommandWrapper.Register("upgrade", (int index) => index == 0 ? Upgrade.Types : FiltererParameters.Parameters, FiltererParameters.GetAutoComplete());
+    new Terminal.ConsoleCommand("upgrade", "[operation] [...args] - Performs a predefined upgrade operation.", (args) =>
+    {
       FiltererParameters pars = new(args);
       var selectedType = "";
-      foreach (var type in Upgrade.Types) {
+      foreach (var type in Upgrade.Types)
+      {
         if (Parse.Flag(pars.Unhandled, type.ToLower())) selectedType = type;
       }
       if (!pars.Valid(args.Context)) return;