+++ UpgradeWorld/commands/Verbose.cs
@@ -1,11 +1,11 @@
-using System.Collections.Generic;
-using System.Linq;
-
 namespace UpgradeWorld;
-public class VerboseCommand {
-  public VerboseCommand() {
+public class VerboseCommand
+{
+  public VerboseCommand()
+  {
     CommandWrapper.RegisterEmpty("verbose");
-    new Terminal.ConsoleCommand("verbose", "- Toggles the verbose mode.", (args) => {
+    new Terminal.ConsoleCommand("verbose", "- Toggles the verbose mode.", (args) =>
+    {
       if (Helper.IsClient(args)) return;
       Settings.configVerbose.Value = !Settings.Verbose;
       if (Settings.Verbose)