+++ UpgradeWorld/commands/ZonesGenerate.cs
@@ -2,13 +2,18 @@ using System;
 using System.Linq;
 
 namespace UpgradeWorld;
-public class ZonesGenerateCommand {
-  public ZonesGenerateCommand() {
-    CommandWrapper.Register("zones_generate", (int index) => {
+public class ZonesGenerateCommand
+{
+  public ZonesGenerateCommand()
+  {
+    CommandWrapper.Register("zones_generate", (int index) =>
+    {
       return FiltererParameters.Parameters;
     }, FiltererParameters.GetAutoComplete());
-    new Terminal.ConsoleCommand("zones_generate", "[...args] - Pre-generates areas without having to visit them.", (args) => {
-      FiltererParameters pars = new(args) {
+    new Terminal.ConsoleCommand("zones_generate", "[...args] - Pre-generates areas without having to visit them.", (args) =>
+    {
+      FiltererParameters pars = new(args)
+      {
         TargetZones = TargetZones.Ungenerated,
         SafeZones = 0
       };