+++ UpgradeWorld/filterers/BiomeFilterer.cs
@@ -3,16 +3,10 @@ using System.Linq;
 using UnityEngine;
 namespace UpgradeWorld;
 ///<summary>Filters zones based on given biomes.</summary>
-public class BiomeFilterer : IZoneFilterer {
-
-  private readonly bool IncludeEdges = true;
-  private readonly IEnumerable<Heightmap.Biome> Biomes;
-  public BiomeFilterer(IEnumerable<Heightmap.Biome> biomes, bool includeEdges) {
-    Biomes = biomes;
-    IncludeEdges = includeEdges;
-  }
-
-  public Vector2i[] FilterZones(Vector2i[] zones, ref List<string> messages) {
+public class BiomeFilterer(IEnumerable<Heightmap.Biome> Biomes, bool IncludeEdges) : IZoneFilterer
+{
+  public Vector2i[] FilterZones(Vector2i[] zones, ref List<string> messages)
+  {
     if (Biomes.Count() == 0) return zones;
     var amount = zones.Length;
     zones = FilterByBiomes(zones);
@@ -21,15 +15,17 @@ public class BiomeFilterer : IZoneFilterer {
     return zones;
   }
 
-  private Vector2i[] FilterByBiomes(Vector2i[] zones) {
+  private Vector2i[] FilterByBiomes(Vector2i[] zones)
+  {
     var zoneSystem = ZoneSystem.instance;
     var halfZone = zoneSystem.m_zoneSize / 2.0f;
-    return zones.Where(zone => {
+    return zones.Where(zone =>
+    {
       var center = zoneSystem.GetZonePos(zone);
-      var corner1 = center + new Vector3(halfZone, 0, halfZone);
-      var corner2 = center + new Vector3(-halfZone, 0, halfZone);
-      var corner3 = center + new Vector3(halfZone, 0, -halfZone);
-      var corner4 = center + new Vector3(-halfZone, 0, -halfZone);
+      var corner1 = center + new Vector3(halfZone, 0f, halfZone);
+      var corner2 = center + new Vector3(-halfZone, 0f, halfZone);
+      var corner3 = center + new Vector3(halfZone, 0f, -halfZone);
+      var corner4 = center + new Vector3(-halfZone, 0f, -halfZone);
       var biome1 = WorldGenerator.instance.GetBiome(corner1);
       var biome2 = WorldGenerator.instance.GetBiome(corner2);
       var biome3 = WorldGenerator.instance.GetBiome(corner3);