+++ UpgradeWorld/filterers/LocationFilterer.cs
@@ -2,20 +2,19 @@ using System.Collections.Generic;
 using System.Linq;
 namespace UpgradeWorld;
 ///<summary>Filters zones based on whether they have given locations.</summary>
-public class LocationFilterer : IZoneFilterer {
-  private readonly IEnumerable<string> Ids;
-  public LocationFilterer(IEnumerable<string> ids) {
-    Ids = ids;
-  }
-  public Vector2i[] FilterZones(Vector2i[] zones, ref List<string> messages) {
+public class LocationFilterer(IEnumerable<string> Ids, bool OnlyMissing) : IZoneFilterer
+{
+  public Vector2i[] FilterZones(Vector2i[] zones, ref List<string> messages)
+  {
     var locationObjects = Ids.Select(id => id.GetStableHashCode()).ToHashSet();
     if (locationObjects.Count == 0) return zones;
     var zs = ZoneSystem.instance;
     var amount = zones.Length;
-    zones = zones.Where(zone => {
-      if (!zs.m_locationInstances.TryGetValue(zone, out var instance)) return false;
-      return locationObjects.Contains(instance.m_location.m_hash);
-    }).ToArray();
+    zones = zones.Where(zone =>
+      zs.m_locationInstances.TryGetValue(zone, out var instance) &&
+      (!OnlyMissing || !instance.m_placed) &&
+      (locationObjects.Count == 0 || locationObjects.Contains(instance.m_location.m_hash))
+    ).ToArray();
     var skipped = amount - zones.Length;
     if (skipped > 0) messages.Add(skipped + " skipped by not having the location");
     return zones;