+++ UpgradeWorld/service/Parse.cs
@@ -6,48 +6,60 @@ using UnityEngine;
 using UpgradeWorld;
 
 namespace Service;
-public class Range<T> where T : IComparable {
+public class Range<T> where T : IComparable
+{
   public T Min;
   public T Max;
-  public Range(T value) {
+  public Range(T value)
+  {
     Min = value;
     Max = value;
   }
-  public Range(T min, T max) {
+  public Range(T min, T max)
+  {
     Min = min;
     Max = max;
   }
 
-  public bool Includes(T value) {
+  public bool Includes(T value)
+  {
     return Min.CompareTo(value) <= 0 && Max.CompareTo(value) >= 0;
   }
 }
-public class Vector3Range {
+public class Vector3Range
+{
   public Vector3 Min;
   public Vector3 Max;
-  public Vector3Range(Vector3 value) {
+  public Vector3Range(Vector3 value)
+  {
     Min = value;
     Max = value;
   }
-  public Vector3Range(Vector3 min, Vector3 max) {
+  public Vector3Range(Vector3 min, Vector3 max)
+  {
     Min = min;
     Max = max;
   }
 
-  public bool Includes(Vector3 value) {
+  public bool Includes(Vector3 value)
+  {
     return Min.sqrMagnitude <= value.sqrMagnitude && value.sqrMagnitude <= Max.sqrMagnitude;
   }
 }
 
 ///<summary>Contains functions for parsing arguments, etc.</summary>
-public static class Parse {
-  private static Range<string> Range(string arg) {
+public static class Parse
+{
+  private static Range<string> Range(string arg)
+  {
     var range = arg.Split('-').ToList();
-    if (range.Count > 1 && range[0] == "") {
+    if (range.Count > 1 && range[0] == "")
+    {
       range[0] = "-" + range[1];
       range.RemoveAt(1);
     }
-    if (range.Count > 2 && range[1] == "") {
+    if (range.Count > 2 && range[1] == "")
+    {
       range[1] = "-" + range[2];
       range.RemoveAt(2);
     }
@@ -55,83 +67,102 @@ public static class Parse {
     else return new(range[0], range[1]);
 
   }
-  public static int Int(string arg, int defaultValue = 0) {
+  public static int Int(string arg, int defaultValue = 0)
+  {
     if (!int.TryParse(arg, NumberStyles.Integer, CultureInfo.InvariantCulture, out var result))
       return defaultValue;
     return result;
   }
-  public static int Int(string[] args, int index, int defaultValue = 0) {
+  public static int Int(string[] args, int index, int defaultValue = 0)
+  {
     if (args.Length <= index) return defaultValue;
     return Int(args[index], defaultValue);
   }
-  public static Range<int> IntRange(string arg, int defaultValue = 0) {
+  public static Range<int> IntRange(string arg, int defaultValue = 0)
+  {
     var range = Range(arg);
     return new(Int(range.Min, defaultValue), Int(range.Max, defaultValue));
   }
-  public static Range<int> IntRange(string[] args, int index, int defaultValue = 0) {
+  public static Range<int> IntRange(string[] args, int index, int defaultValue = 0)
+  {
     if (args.Length <= index) return new(defaultValue);
     return IntRange(args[index], defaultValue);
   }
-  public static uint UInt(string arg, uint defaultValue = 0) {
+  public static uint UInt(string arg, uint defaultValue = 0)
+  {
     if (!uint.TryParse(arg, NumberStyles.Integer, CultureInfo.InvariantCulture, out var result))
       return defaultValue;
     return result;
   }
-  public static uint UInt(string[] args, int index, uint defaultValue = 0) {
+  public static uint UInt(string[] args, int index, uint defaultValue = 0)
+  {
     if (args.Length <= index) return defaultValue;
     return UInt(args[index], defaultValue);
   }
-  public static Range<uint> UIntRange(string arg, uint defaultValue = 0) {
+  public static Range<uint> UIntRange(string arg, uint defaultValue = 0)
+  {
     var range = Range(arg);
     return new(UInt(range.Min, defaultValue), UInt(range.Max, defaultValue));
   }
-  public static Range<uint> UIntRange(string[] args, int index, uint defaultValue = 0) {
+  public static Range<uint> UIntRange(string[] args, int index, uint defaultValue = 0)
+  {
     if (args.Length <= index) return new(defaultValue);
     return UIntRange(args[index], defaultValue);
   }
-  public static long Long(string arg, long defaultValue = 0) {
+  public static long Long(string arg, long defaultValue = 0)
+  {
     if (!long.TryParse(arg, NumberStyles.Integer, CultureInfo.InvariantCulture, out var result))
       return defaultValue;
     return result;
   }
-  public static long Long(string[] args, int index, long defaultValue = 0) {
+  public static long Long(string[] args, int index, long defaultValue = 0)
+  {
     if (args.Length <= index) return defaultValue;
     return Long(args[index], defaultValue);
   }
-  public static Range<long> LongRange(string arg, long defaultValue = 0) {
+  public static Range<long> LongRange(string arg, long defaultValue = 0)
+  {
     var range = Range(arg);
     return new(Long(range.Min, defaultValue), Long(range.Max, defaultValue));
   }
-  public static Range<long> LongRange(string[] args, int index, long defaultValue = 0) {
+  public static Range<long> LongRange(string[] args, int index, long defaultValue = 0)
+  {
     if (args.Length <= index) return new(defaultValue);
     return LongRange(args[index], defaultValue);
   }
-  public static bool TryFloat(string arg, out float value) {
+  public static bool TryFloat(string arg, out float value)
+  {
     return float.TryParse(arg, NumberStyles.Float, CultureInfo.InvariantCulture, out value);
   }
-  public static float Float(string arg, float defaultValue = 0f) {
+  public static float Float(string arg, float defaultValue = 0f)
+  {
     if (!TryFloat(arg, out var result))
       return defaultValue;
     return result;
   }
-  public static float Float(string[] args, int index, float defaultValue = 0f) {
+  public static float Float(string[] args, int index, float defaultValue = 0f)
+  {
     if (args.Length <= index) return defaultValue;
     return Float(args[index], defaultValue);
   }
-  public static Range<float> FloatRange(string arg, float defaultValue = 0f) {
+  public static Range<float> FloatRange(string arg, float defaultValue = 0f)
+  {
     var range = Range(arg);
     return new(Float(range.Min, defaultValue), Float(range.Max, defaultValue));
   }
-  public static Range<float> FloatRange(string[] args, int index, float defaultValue = 0f) {
+  public static Range<float> FloatRange(string[] args, int index, float defaultValue = 0f)
+  {
     if (args.Length <= index) return new(defaultValue);
     return FloatRange(args[index], defaultValue);
   }
-  public static string String(string[] args, int index, string defaultValue = "") {
+  public static string String(string[] args, int index, string defaultValue = "")
+  {
     if (args.Length <= index) return defaultValue;
     return args[index];
   }
   public static Quaternion AngleYXZ(string arg) => AngleYXZ(arg, Quaternion.identity);
-  public static Quaternion AngleYXZ(string arg, Quaternion defaultValue) {
+  public static Quaternion AngleYXZ(string arg, Quaternion defaultValue)
+  {
     var values = Split(arg);
     var angle = Vector3.zero;
     angle.y = Parse.Float(values, 0, defaultValue.eulerAngles.y);
@@ -147,14 +178,16 @@ public static class Parse {
   ///<summary>Parses XZY vector starting at given index. Zero is used for missing values.</summary>
   public static Vector3 VectorXZY(string[] args, int index) => VectorXZY(args, index, Vector3.zero);
   ///<summary>Parses XZY vector starting at given index. Default values is used for missing values.</summary>
-  public static Vector3 VectorXZY(string[] args, int index, Vector3 defaultValue) {
+  public static Vector3 VectorXZY(string[] args, int index, Vector3 defaultValue)
+  {
     var vector = Vector3.zero;
     vector.x = Float(args, index, defaultValue.x);
     vector.y = Float(args, index + 2, defaultValue.y);
     vector.z = Float(args, index + 1, defaultValue.z);
     return vector;
   }
-  public static Vector3Range VectorXZYRange(string arg, Vector3 defaultValue) {
+  public static Vector3Range VectorXZYRange(string arg, Vector3 defaultValue)
+  {
     var parts = Split(arg);
     var x = FloatRange(parts, 0, defaultValue.x);
     var y = FloatRange(parts, 2, defaultValue.y);
@@ -168,21 +201,24 @@ public static class Parse {
   ///<summary>Parses ZXY vector starting at given index. Zero is used for missing values.</summary>
   public static Vector3 VectorZXY(string[] args, int index) => VectorZXY(args, index, Vector3.zero);
   ///<summary>Parses ZXY vector starting at given index. Default values is used for missing values.</summary>
-  public static Vector3 VectorZXY(string[] args, int index, Vector3 defaultValue) {
+  public static Vector3 VectorZXY(string[] args, int index, Vector3 defaultValue)
+  {
     var vector = Vector3.zero;
     vector.x = Float(args, index + 1, defaultValue.x);
     vector.y = Float(args, index + 2, defaultValue.y);
     vector.z = Float(args, index, defaultValue.z);
     return vector;
   }
-  public static Vector3Range VectorZXYRange(string arg, Vector3 defaultValue) {
+  public static Vector3Range VectorZXYRange(string arg, Vector3 defaultValue)
+  {
     var parts = Split(arg);
     var x = FloatRange(parts, 1, defaultValue.x);
     var y = FloatRange(parts, 2, defaultValue.y);
     var z = FloatRange(parts, 0, defaultValue.z);
     return ToVectorRange(x, y, z);
   }
-  private static Vector3Range ToVectorRange(Range<float> x, Range<float> y, Range<float> z) {
+  private static Vector3Range ToVectorRange(Range<float> x, Range<float> y, Range<float> z)
+  {
     Vector3 min = new(x.Min, y.Min, z.Min);
     Vector3 max = new(x.Max, y.Max, z.Max);
     return new(min, max);
@@ -194,14 +230,16 @@ public static class Parse {
   ///<summary>Parses YXZ vector starting at given index. Zero is used for missing values.</summary>
   public static Vector3 VectorYXZ(string[] args, int index) => VectorYXZ(args, index, Vector3.zero);
   ///<summary>Parses YXZ vector starting at given index. Default values is used for missing values.</summary>
-  public static Vector3 VectorYXZ(string[] args, int index, Vector3 defaultValue) {
+  public static Vector3 VectorYXZ(string[] args, int index, Vector3 defaultValue)
+  {
     var vector = Vector3.zero;
     vector.y = Float(args, index, defaultValue.y);
     vector.x = Float(args, index + 1, defaultValue.x);
     vector.z = Float(args, index + 2, defaultValue.z);
     return vector;
   }
-  public static Vector3Range VectorYXZRange(string arg, Vector3 defaultValue) {
+  public static Vector3Range VectorYXZRange(string arg, Vector3 defaultValue)
+  {
     var parts = Split(arg);
     var x = FloatRange(parts, 1, defaultValue.x);
     var y = FloatRange(parts, 0, defaultValue.y);
@@ -212,14 +250,16 @@ public static class Parse {
   public static Vector3 Scale(string[] args) => Scale(args, 0);
   ///<summary>Parses scale starting at given index. Includes a sanity check and giving a single value for all axis.</summary>
   public static Vector3 Scale(string[] args, int index) => SanityCheck(VectorXZY(args, index));
-  private static Vector3 SanityCheck(Vector3 scale) {
+  private static Vector3 SanityCheck(Vector3 scale)
+  {
     // Sanity check and also adds support for setting all values with a single number.
     if (scale.x == 0) scale.x = 1;
     if (scale.y == 0) scale.y = scale.x;
     if (scale.z == 0) scale.z = scale.x;
     return scale;
   }
-  public static Vector3Range ScaleRange(string arg) {
+  public static Vector3Range ScaleRange(string arg)
+  {
     var parts = Split(arg);
     var x = FloatRange(parts, 0, 0f);
     var y = FloatRange(parts, 1, 0f);
@@ -231,10 +271,7 @@ public static class Parse {
   }
 
   public static string[] Split(string arg, char separator = ',') => arg.Split(separator).Select(s => s.Trim()).Where(s => s != "").ToArray();
-  public static string[] Split(string[] args, int index, char separator) {
-    if (args.Length <= index) return new string[0];
-    return Split(args[index], separator);
-  }
+  public static string[] Split(string[] args, int index, char separator) => args.Length <= index ? ([]) : Split(args[index], separator);
   private static readonly HashSet<string> Truthies = new() {
     "1",
     "true",
@@ -249,29 +286,34 @@ public static class Parse {
     "off"
   };
   private static bool IsFalsy(string value) => Falsies.Contains(value);
-  public static bool? Boolean(string arg) {
+  public static bool? Boolean(string arg)
+  {
     if (IsTruthy(arg)) return true;
     if (IsFalsy(arg)) return false;
     return null;
   }
-  public static IEnumerable<string> Flag(IEnumerable<string> parameters, string flag, out bool value) {
+  public static IEnumerable<string> Flag(IEnumerable<string> parameters, string flag, out bool value)
+  {
     value = parameters.FirstOrDefault(arg => arg.ToLower() == flag) != null;
     return parameters.Where(arg => arg.ToLower() != flag);
   }
-  public static bool Flag(List<string> parameters, string flag) {
+  public static bool Flag(List<string> parameters, string flag)
+  {
     var key = parameters.FirstOrDefault(value => value.ToLower() == flag.ToLower());
     if (key == null) return false;
     return parameters.Remove(key);
   }
 
-  public static Vector2i Zone(string arg) {
+  public static Vector2i Zone(string arg)
+  {
     var values = Split(arg).ToArray();
     Vector2i vector = new();
     if (values.Length > 0) vector.x = Int(values[0]);
     if (values.Length > 1) vector.y = Int(values[1]);
     return vector;
   }
-  public static Vector2 Pos(string arg) {
+  public static Vector2 Pos(string arg)
+  {
     var values = Split(arg).ToArray();
     Vector2 vector = new();
     if (values.Length > 0) vector.x = Float(values[0]);