+++ publish/CHANGELOG.md
@@ -1,3 +1,7 @@
+- v1.41
+  - Improves the performance `locations_add` command (used by most upgrade commands).
+  - Fixes `locations_add` command placing every unplaced location, instead of just the specified locations (mostly affects some custom maps).
+
 - v1.40
   - Adds a new upgrade `hildir`.
   - Fixes for the new update.